# openapi-rest.yaml
openapi: 3.0.0
info:
  description: API server for Hyperledger Fabric
  title: fabric-starter-rest
  version: FABRIC_STARTER_REST_VERSION=latest
paths:
  /users:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        description: User object
        required: true
      description: |-

        Login or register user.
      tags:
        - users
      responses:
        "200":
          description: User logged in and his JWT returned
        default:
          description: Unexpected error
  /domain:
    get:
      description: |-

        Show network name (as defined by DOMAIN env variable at setup time)
      tags:
        - config
      responses:
        "200":
          description: DOMAIN
        default:
          description: Unexpected error
  /mspid:
    get:
      description: |-

        Show name (MSPID) of my organization
      tags:
        - config
      responses:
        "200":
          description: MSPID
        default:
          description: Unexpected error
  /config:
    get:
      description: >-

        Network config json to aid debugging; use for development only as it may expose sensitive data
      tags:
        - config
      responses:
        "200":
          description: Network config
        default:
          description: Unexpected error
...
servers:
  - url: /
components:
...
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: Authorization
      description: "Paste the jwt you received from logging in by a post to /users ex.:
        Bearer eyJhbGci..."
  schemas:
    User:
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: "username "
          example: oleg
        password:
          type: string
          description: "password "
          example: pass
...